pagamento-service/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy-Utils
pymysql
python-dotenv
pydantic[email]